Revert "Use Duplex instead of Sink for pw-loopback"
This reverts commit 79f6bf8da1bc8dd87bc0bc5256f240fcf3a00353.

Got the taste of issues I forgot. The sink unmutes automatically
for some reason (maybe on window change?) which is annoying. And
that is after manually muting from pavucontrol.

Also, EasyEffects is unable to detect Broo.

Of course, this revert affects applications which are incorrectly
coded and do not detect Monitor sources, like Zoom. It probably
wants a dedicated source.