broo: Remap sink to a dedicated source.
Let's create a dedicated source which would be used as the input mic by
the programs. We will name it Broo, and name the sink as Shru to avoid
confusion.

We do this because Monitor sources are not usable by many apps like
GNOME settings or Zoom, they expect dedicated sources.

Fixes #24